Fixed preloading of data
