+++ posthog-react-native/index.ts
@@ -5,4 +5,6 @@ export * from './src/posthog-rn'
 export * from './src/hooks/useLifecycleTracker'
 export * from './src/hooks/useNavigationTracker'
 export * from './src/hooks/useFeatureFlags'
+export * from './src/hooks/useFeatureFlag'
+export * from './src/hooks/usePostHog'
 export * from './src/PostHogProvider'
