+++ posthog-react-native/src/PosthogContext.ts
@@ -1,0 +1,4 @@
+import React from 'react'
+import { PostHog } from './posthog-rn'
+
+export const PostHogContext = React.createContext<{ client?: PostHog }>({ client: undefined })
